backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
fastapi-jwt-auth
requests